agent_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
langchain
langchain-nvidia-ai-endpoints
pymupdf